Replace PyTorch with ONNX Runtime + FP16 model for Render free tier
- Export U-Net as FP16 ONNX (46.6 MB, weights embedded)
- Replace torch/torchvision with onnxruntime (30 MB vs 300 MB)
- Lazy-load model on first request
- Memory: ~80 MB total vs 512 MB limit

+++ model_utils.py
@@ -92,7 +92,7 @@ def preprocess_image(image_bytes: bytes) -> np.ndarray:
 
     # HWC -> CHW -> (1, C, H, W)
     img = np.transpose(img, (2, 0, 1))
-    return img[np.newaxis, ...].astype(np.float32)
+    return img[np.newaxis, ...].astype(np.float16)
 
 
 def run_inference(session, tensor: np.ndarray) -> tuple[np.ndarray, np.ndarray]:
